+++ src/main/java/com/project_management/shoppingweb/repository/ShoppingCartRepository.java
@@ -9,5 +9,5 @@ import java.util.List;
 @Repository
 public interface ShoppingCartRepository extends JpaRepository<ShoppingCart,Long>{
     ShoppingCart save(ShoppingCart shoppingCart);
-    List<ShoppingCart> findByUserId(int usrId);
+    List<ShoppingCart> findByUserId(Long usrId);
 }
