+++ src/main/java/com/project_management/shoppingweb/repository/TradeDetailRepository.java
@@ -9,6 +9,6 @@ import java.util.List;
 @Repository
 public interface TradeDetailRepository extends JpaRepository<TradeDetail,Long>{
     TradeDetail save(TradeDetail tradeDetail);
-    List<TradeDetail> findByTradeId(int TradeId);
-    TradeDetail findByTradeDetailId(int tradeDetailId);
+    List<TradeDetail> findByTradeId(Long TradeId);
+    TradeDetail findByTradeDetailId(Long tradeDetailId);
 }
